requirements.txt
fastapi
uvicorn
pytesseract
pillow
dateparser
python-multipart
